refactor: компонент ComponentRemove переименован на ComponentRemove
