+++ packages/core/components/component-remove/index.ts
@@ -1,0 +1,2 @@
+export { ComponentRemove } from './component-remove';
+export type { Props as ComponentsRemoveProps } from './component-remove';
