--- packages/core/components/components-remove/index.ts
@@ -1,2 +1,0 @@
-export { ComponentsRemove } from './components-remove';
-export type { Props as ComponentsRemoveProps } from './components-remove';
